backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.28.1
openai==1.58.1
transformers==4.46.3
torch==2.5.1
whisper==1.1.10
redis==5.2.1
celery==5.4.0
websockets==14.1
aiofiles==24.1.0
Pillow==11.1.0
pypdf==5.1.0
nltk==3.9.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
pytest==8.3.4
pytest-asyncio==0.25.0